Implementation of shader generator for Splat Materials supporting gamma and tonemapping options
